refactored configuration route
