+++ src/app/api/instances/[name]/logs/utils/getLogs.ts
@@ -1,9 +1,9 @@
 import { fetchInstance } from "@/utils/AWS/EC2/fetchInstance";
 import { runSSMCommands } from "@/utils/AWS/SSM/runSSMCommands";
 import { EC2Client } from "@aws-sdk/client-ec2";
-import { GetLogsParams } from "./types";
+import { GetLogsParams } from "../types";
 
-export async function getLogs({
+export default async function getLogs({
   region,
   instanceName,
 }: GetLogsParams): Promise<string> {
