+++ src/app/api/instances/[name]/plugins/route.ts
@@ -1,7 +1,8 @@
 import { NextRequest, NextResponse } from "next/server";
 
-import { getPlugins, togglePlugins } from "./service";
-
+// import { getPlugins, togglePlugins } from "./service";
+import getPlugins from "./utils/getPlugins";
+import togglePlugins from "./utils/togglePlugins";
 import eventEmitter from "@/utils/eventEmitter";
 import { deleteEvent } from "@/utils/eventBackups";
